requirements.txt
yfinance
kafka-python
pandas
pyarrow
pyspark
numpy
matplotlib
seaborn
scikit-learn